+++ src/main/java/com/lowdragmc/mbd2/common/machine/definition/config/event/MachineCustomDataUpdateEvent.java
@@ -5,12 +5,10 @@ import com.lowdragmc.lowdraglib.gui.graphprocessor.data.parameter.ExposedParamet
 import com.lowdragmc.mbd2.common.machine.MBDMachine;
 import com.lowdragmc.mbd2.common.graphprocessor.GraphParameterGet;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraftforge.eventbus.api.Cancelable;
 
 import java.util.Map;
 import java.util.Optional;
 
-@Cancelable
 @LDLRegister(name = "MachineCustomDataUpdateEvent", group = "MachineEvent")
 public class MachineCustomDataUpdateEvent extends MachineEvent {
     @GraphParameterGet(displayName = "old data")