+++ src/main/java/com/lowdragmc/mbd2/integration/kubejs/MBDKubeJSPlugin.java
@@ -6,10 +6,7 @@ import com.lowdragmc.mbd2.api.registry.MBDRegistries;
 import com.lowdragmc.mbd2.common.machine.definition.config.ConfigBlockProperties;
 import com.lowdragmc.mbd2.common.machine.definition.config.MachineState;
 import com.lowdragmc.mbd2.integration.create.machine.CreateMachineState;
-import com.lowdragmc.mbd2.integration.kubejs.events.MBDClientEvents;
-import com.lowdragmc.mbd2.integration.kubejs.events.MBDMachineEvents;
-import com.lowdragmc.mbd2.integration.kubejs.events.MBDServerEvents;
-import com.lowdragmc.mbd2.integration.kubejs.events.MBDStartupEvents;
+import com.lowdragmc.mbd2.integration.kubejs.events.*;
 import com.lowdragmc.mbd2.integration.kubejs.recipe.MBDRecipeSchema;
 import dev.latvian.mods.kubejs.KubeJSPlugin;
 import dev.latvian.mods.kubejs.recipe.RecipesEventJS;
@@ -42,6 +39,7 @@ public class MBDKubeJSPlugin extends KubeJSPlugin {
         }
         MBDStartupEvents.REGISTRY_EVENTS.register();
         MBDMachineEvents.MBD_MACHINE_EVENTS.register();
+        MBDRecipeTypeEvents.MBD_RECIPE_TYPE_EVENTS.register();
     }
 
     @Override