+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * Bump up photon version, and fix photon version compatibility
 * Fixed After Recipe Working Event also triggers when structure is destroyed
 * Added try-catch to prevent crash when trying to execute kjs scripts
+* Added TransferProxyRecipeEvent to allow customizing transfer proxy recipe behavior
+* Added MBDRecipeType#recipeBuilder to build dynamic recipe as kjs recipe event